renamed modules/files/packages named mcp
