+++ apps/mcp_server/mcp_server/tools/get_session_info.py
