+++ lib/config/schema/__init__.py
@@ -31,7 +31,7 @@ from .hud import (INPUT_TELEMETRY_OVERLAY_ID, LAP_TIMER_OVERLAY_ID,
                   TRACK_MAP_OVERLAY_ID, TRACK_RADAR_OVERLAY_ID, HudSettings,
                   MfdPageSettings, MfdSettings, OverlayPosition,
                   TimingTowerColOptions)
-from .mcp import McpSettings
+from .mcp_cfg import McpSettings
 from .network import NetworkSettings
 from .pit_time_loss_f1 import PitTimeLossF1
 from .pit_time_loss_f2 import PitTimeLossF2
