+++ lib/config/schema/mcp_cfg.py
